requirements.txt
pandas
python-dotenv
requests
openpyxl
